Refactored so no commands were added to extension.ts
